Updated files (git_understand : JAVA, C++ ; git_understand_python : Python ; git_understand_non_obj : Fortran

+++ github/API_V3/git_understand/git_understand.py
@@ -5,7 +5,15 @@ Created on Tue Mar  5 18:54:29 2019
 
 @author: suvodeepmajumder
 """
+import codecs
+import sys
+import subprocess
+import shlex
+import json  # json is an easy way to send arbitrary ascii-safe lists of strings out of python
+import sys
+sys.path.append("..")
 from git_log import git2repo
+#from main.git_log import git2repo
 import os
 import re
 import shlex
@@ -23,6 +31,25 @@ from utils.utils import utils
 import platform
 from os.path import dirname as up
 #from main.utils.utils.utils import printProgressBar
+
+
+def shell_split(cmd):
+    if os.name == 'posix':
+        return shlex.split(cmd)
+    else:
+        # TODO: write a version of this that doesn't invoke a subprocess
+        if not cmd:
+            return []
+        #cmd = cmd.replace("\\\\", "\\")
+        print("print cmd : ",cmd)
+        full_cmd = '{} {}'.format(
+            subprocess.list2cmdline([
+                sys.executable, '-c',
+                'import sys, json; print(json.dumps(sys.argv[1:]))'
+            ]), cmd
+        )
+        ret = subprocess.check_output(full_cmd).decode()
+        return json.loads(ret)
 
 
 class MetricsGetter(object):
@@ -46,13 +73,14 @@ class MetricsGetter(object):
         self.repo = self.repo_obj.clone_repo()
         if platform.system() == 'Darwin' or platform.system() == 'Linux':
             self.repo_path = up(os.getcwd()) + '/temp_repo/' + self.repo_name
-            self.file_path = up(os.getcwd()) + '/data/commit/' + self.repo_name + '_commit.pkl'
-            self.committed_file = up(os.getcwd()) + '/data/committed_files/' + self.repo_name + '_committed_file.pkl'
+            self.file_path = up(os.getcwd()) + '/data/' + self.repo_name + '_commit.pkl'
+            self.committed_file = up(os.getcwd()) + '/data/' + self.repo_name + '_committed_file.pkl'
         else:
             self.repo_path = up(os.getcwd()) + '\\temp_repo\\' + self.repo_name
-            self.file_path = up(os.getcwd()) + '\\data\\commit\\' + self.repo_name + '_commit.pkl'
-            self.committed_file = up(os.getcwd()) + '\\data\\committed_files\\' + self.repo_name + '_committed_file.pkl'
+            self.file_path = up(os.getcwd()) + '\\data\\' + self.repo_name + '_commit.pkl'
+            self.committed_file = up(os.getcwd()) + '\\data\\' + self.repo_name + '_committed_file.pkl'
         self.buggy_clean_pairs = self.read_commits()
+        print("generated buggy clean pairs")
         self.repo_path = self.repo_obj.repo_path
         # Reference current directory, so we can go back after we are done.
         self.cwd = Path(os.getcwd())
@@ -67,32 +95,41 @@ class MetricsGetter(object):
         # Generate source path where the source file exist
         self.source_path = self.cwd.joinpath(
             ".temp", "sources", self.repo_name)
+        print(self.source_path)
 
 
     def read_commits(self):
+        print(self.file_path)
         df = pd.read_pickle(self.file_path)
         df_committed_file = pd.read_pickle(self.committed_file)
         df = df[df['buggy'] == 1]
         df_commits = df.drop(labels = ['message','buggy'], axis = 1)
         commits = []
+        temp = df_commits.shape[0]
+        print("here are they :",temp)
         for i in range(df_commits.shape[0]):
             committed_files = []
             if df_commits.iloc[i,0] == None or df_commits.iloc[i,1] == None:
                 continue
+            #print("printing row of commits : \n", df_commits.iloc[i,1])
             bug_fixing_commit = self.repo.get(df_commits.iloc[i,0])
             bug_existing_commit = self.repo.get(df_commits.iloc[i,1])
             if bug_fixing_commit == None:
                 print(df_commits.iloc[i,0])
                 continue
             for index,row in  df_committed_file[df_committed_file['commit_id'] == df_commits.iloc[i,0]].iterrows():
+                a = row['file_path']
                 if len(row['file_path'].split('src/')) == 1:
+                    print("3")
                     continue
+                temp = row['file_path']
+                b = temp.split('src/')[1].replace('/', '.').rsplit('.', 1)[0]
                 committed_files.append(row['file_path'].split('src/')[1].replace('/','.').rsplit('.',1)[0])
             commits.append([bug_existing_commit,bug_fixing_commit,committed_files])
         return commits
 
     @staticmethod
-    def _os_cmd(cmd, verbose=False):
+    def _os_cmd(cmd, verbose=True):
         """
         Run a command on the shell
 
@@ -101,8 +138,9 @@ class MetricsGetter(object):
         cmd: str
             A command to run.
         """
-        cmd = shlex.split(cmd)
-        #print(cmd)
+        #print("string : ", cmd)
+        cmd = shell_split(cmd)
+        #cmd = shlex.split(cmd)
         with sp.Popen(cmd, stdout=sp.PIPE, stderr=sp.DEVNULL) as p:
             out, err = p.communicate()
 
@@ -133,16 +171,16 @@ class MetricsGetter(object):
 
         # Generate udb file
         if self.repo_lang == "fortran":
-            cmd = "/Applications/Understand.app/Contents/MacOS/und create -languages Fortran add {} analyze {}".format(
+            cmd = "und create -languages Fortran add {} analyze {}".format(
                 str(self.repo_path), str(und_file))
         elif self.repo_lang == "python":
-            cmd = "/Applications/Understand.app/Contents/MacOS/und create -languages python add {} analyze {}".format(
+            cmd = "und create -languages python add {} analyze {}".format(
                 str(self.repo_path), str(und_file))
         elif self.repo_lang == "C":
-            cmd = "/Applications/Understand.app/Contents/MacOS/und create -languages C++ add {} analyze {}".format(
+            cmd = "und create -languages C++ add {} analyze {}".format(
                 str(self.repo_path), str(und_file))
         elif self.repo_lang == "java":
-            cmd = "/Applications/Understand.app/Contents/MacOS/und create -languages Java add {} analyze {}".format(
+            cmd = "und create -languages Java add {} analyze {}".format(
                 str(self.repo_path), str(und_file))
         out, err = self._os_cmd(cmd)
 
@@ -204,14 +242,22 @@ class MetricsGetter(object):
         """
 
         self.metrics_dataframe = pd.DataFrame()
+
+        #printProgressBar(0, len(self.buggy_clean_pairs), prefix='Progress:',
+        #                 suffix='Complete', length=50)
+
+        # 1. For each clean-buggy commit pairs
         print(len(self.buggy_clean_pairs))
         for i in range(len(self.buggy_clean_pairs)):
             buggy_hash = self.buggy_clean_pairs[i][0]
             clean_hash = self.buggy_clean_pairs[i][1]
             files_changed = self.buggy_clean_pairs[i][2]
+            if len(files_changed) == 0:
+                continue
             print(i,(buggy_hash.id.hex, clean_hash.id.hex))
             # Go the the cloned project path
             os.chdir(self.repo_path)
+            #print(self.repo_path)
             # Checkout the master branch first, we'll need this
             # to find what files have changed.
             self._os_cmd("git reset --hard master", verbose=False)
@@ -230,23 +276,31 @@ class MetricsGetter(object):
             self._create_und_files("buggy")
             #print(self.buggy_und_file)
             db_buggy = und.open(str(self.buggy_und_file))
-            #print("Files",set(files_changed))
+            temp = db_buggy.ents("Class")
             for file in db_buggy.ents("Class"):
                 # print directory name
-                #print(file,file.longname(), file.kind())
                 r = re.compile(str(file.longname()))
                 newlist = list(filter(r.search, list(set(files_changed))))
-                #print(newlist)
                 if len(newlist) > 0:
                     metrics = file.metric(file.metrics())
                     metrics["Name"] = file.longname()
                     metrics["Bugs"] = 1
                     self.metrics_dataframe = self.metrics_dataframe.append(
                         pd.Series(metrics), ignore_index=True)
+                '''
+                if str(file) in files_changed:
+                    print(str(file))
+                    metrics = file.metric(file.metrics())
+                    metrics["Name"] = file.name()
+                    metrics["Bugs"] = 1
+                    self.metrics_dataframe = self.metrics_dataframe.append(
+                        pd.Series(metrics), ignore_index=True)
+                '''
             # Purge und file
             db_buggy.close()
-            #break
-            self._os_cmd("rm {}".format(str(self.buggy_und_file)))
+            #self._os_cmd("rm {}".format(str(self.buggy_und_file)))
+            os.remove(str(self.buggy_und_file))
+
             # ------------------------------------------------------------------
             # ---------------------- CLEAN FILES METRICS -----------------------
             # ------------------------------------------------------------------
@@ -256,23 +310,33 @@ class MetricsGetter(object):
 
             # Create a understand file for this hash
             self._create_und_files("clean")
+            print(files_changed)
             db_clean = und.open(str(self.clean_und_file))
             for file in db_clean.ents("class"):
                 # print directory name
                 r = re.compile(str(file.longname()))
                 newlist = list(filter(r.search, files_changed))
-                #if str(file) in files_changed:
                 if len(newlist) > 0:
                     metrics = file.metric(file.metrics())
-                    #print(metrics)
+                    # print(metrics)
                     metrics["Name"] = file.name()
                     metrics["Bugs"] = 0
                     self.metrics_dataframe = self.metrics_dataframe.append(
                         pd.Series(metrics), ignore_index=True)
+                '''
+                if str(file) in files_changed:
+                    print(str(file))
+                    metrics = file.metric(file.metrics())
+                    metrics["Name"] = file.name()
+                    metrics["Bugs"] = 0
+                    self.metrics_dataframe = self.metrics_dataframe.append(
+                        pd.Series(metrics), ignore_index=True)
+                '''
             db_clean.close()
             # Purge und file
-            self._os_cmd("rm {}".format(str(self.clean_und_file)))
-            #print(self.metrics_dataframe)
+            #self._os_cmd("rm {}".format(str(self.clean_und_file)))
+            os.remove(str(self.clean_und_file))
+            print(self.metrics_dataframe)
 
             #printProgressBar(i, len(self.buggy_clean_pairs),
             #                 prefix='Progress:', suffix='Complete', length=50)
@@ -317,3 +381,17 @@ class MetricsGetter(object):
         self._os_cmd("rm -rf {}/*und".format(self.udb_path))
         # Optional -- remove the clone repo to save some space.
         # self._os_cmd("rm -rf {}".format(self.source_path))
+
+#obj = MetricsGetter("https://github.com/shiffman/Most-Pixels-Ever-Processing","Most-Pixels-Ever-Processing", "java")
+#obj = MetricsGetter("https://github.com/jenkinsci/subversion-plugin", "subversion-plugin", "java")
+#obj = MetricsGetter("https://github.com/spring-projects/spring-retry","spring-retry", "java")
+#obj = MetricsGetter("https://github.com/asterisk-java/asterisk-java","asterisk-java", "java")
+#obj = MetricsGetter("https://github.com/taksan/skype-java-api","skype-java-api", "java")
+#obj = MetricsGetter("https://github.com/onelogin/java-saml","java-saml", "java")
+#obj = MetricsGetter("https://github.com/xmrig/xmrig", "xmrig", "C")
+#obj = MetricsGetter("https://github.com/ThrowTheSwitch/Unity", "Unity", "C")
+#obj = MetricsGetter("https://github.com/libjpeg-turbo/libjpeg-turbo", "libjpeg-turbo", "C")
+obj = MetricsGetter("https://github.com/openresty/lua-nginx-module", "lua-nginx-module", "C")
+ans = obj.get_all_metrics()
+print("Done with metrics")
+
